make tags modular;
