+++ src/components/experiments/Experiment.tsx
@@ -7,6 +7,7 @@ import { Montserrat } from "next/font/google";
 import { ExperimentData, Votes } from "@/types";
 import { Spinner } from "../icons/svgIcons";
 import TagsModal from "./TagsModal";
+import Tags from "./Tags";
 
 const montserrat = Montserrat({ subsets: ["latin"] });
 
@@ -21,30 +22,6 @@ const Experiment = () => {
   const [choice, setChoice] = useState("");
   const [votes, setVotes] = useState<Votes>({} as Votes);
   const [hasPastVote, setHasPastVote] = useState(false);
-  const [showTags, setShowTags] = useState(false);
-  const [isOverflow, setIsOverflow] = useState(false);
-  const tagsContainerRef = useRef<HTMLDivElement>(null);
-
-  // If tags overflow, display "more"
-  useEffect(() => {
-    const container = tagsContainerRef.current;
-    if (!container) return;
-
-    // Recalculate on window resize
-    const observer = new ResizeObserver(() => {
-      if (container) {
-        setIsOverflow(container.scrollWidth > container.clientWidth);
-      }
-    });
-
-    observer.observe(container);
-
-    return () => {
-      if (container) {
-        observer.unobserve(container);
-      }
-    };
-  }, [experiment.tags]);
 
   // Respond to experiment slug change
   useEffect(() => {
@@ -105,10 +82,6 @@ const Experiment = () => {
       <div className="h-[65vh] flex justify-center items-center">{error}</div>
     );
   }
-
-  const handleMoreTagsClick = () => {
-    setShowTags(true);
-  };
 
   // Record user vote & get latest vote tallies
   async function handleVote(vote: string) {
@@ -185,31 +158,7 @@ const Experiment = () => {
             : {experiment.origin}
           </p>
           <div className="w-full mt-2 flex gap-1">
-            <div
-              className={`w-full flex flex-nowrap gap-1 overflow-hidden ${
-                isOverflow ? "flex-grow" : "justify-center"
-              }`}
-              ref={tagsContainerRef}
-            >
-              {experiment.tags.map((tag, index) => {
-                return (
-                  <div
-                    key={index}
-                    className="py-[2px] px-[5px] text-xs whitespace-nowrap rounded text-white dark:text-neutral-200 bg-pool-600 dark:bg-pool-900"
-                  >
-                    {tag}
-                  </div>
-                );
-              })}
-            </div>
-            {isOverflow && (
-              <div
-                className="flex-shrink-0 py-[2px] px-[5px] text-xs rounded text-pool-600 dark:text-pool-500 hover:text-pool-300 whitespace-nowrap cursor-pointer font-bold"
-                onClick={handleMoreTagsClick}
-              >
-                ...more
-              </div>
-            )}
+            <Tags tags={experiment.tags} />
           </div>
         </div>
 
@@ -291,7 +240,6 @@ const Experiment = () => {
           )}
         </div>
       </div>
-      {showTags && <TagsModal setIsOpen={setShowTags} tags={experiment.tags} />}
     </div>
   );
 };
